+++ tests/testthat/test-dbpedia.R
@@ -1,8 +1,9 @@
 test_that(
   "get_dbpedia_uris() for quanteda corpus",
   {
-    library(quanteda)
-    
+
+    withr::local_package("quanteda")
+
     uritab <- quanteda::data_char_ukimmig2010[1:3] %>%
       corpus() %>%
       get_dbpedia_uris(